Basic command prompt running
Using member functions in the 'Console'
class within a map from string to handler.

I thinking each function should just handle
'?' as an argument for printing detailed
help on a given function. The top level
help function '?' will just print the names
of all the commands.

downside of the map is I can't enforce
any order in the help table.